requirements.txt
# Post-Docling Pipeline Requirements
# Core dependencies
guidance>=0.1.15
google-generativeai>=0.8.0
google-cloud-storage>=2.17.0
google-cloud-logging>=3.10.0

# Airflow for orchestration
apache-airflow>=2.8.0
apache-airflow-providers-google>=10.18.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.0
click>=8.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0

# Optional: For enhanced JSON processing
orjson>=3.9.0
jsonschema>=4.21.0
